+++ libyul/optimiser/ASTWalker.h
@@ -56,6 +56,7 @@ public:
 	virtual void operator()(ForLoop const&);
 	virtual void operator()(Break const&) {}
 	virtual void operator()(Continue const&) {}
+	virtual void operator()(Leave const&) {}
 	virtual void operator()(Block const& _block);
 
 	virtual void visit(Statement const& _st);
@@ -91,6 +92,7 @@ public:
 	virtual void operator()(ForLoop&);
 	virtual void operator()(Break&);
 	virtual void operator()(Continue&);
+	virtual void operator()(Leave&);
 	virtual void operator()(Block& _block);
 
 	virtual void visit(Statement& _st);